Moved StringUtils and UrlUtils to com.ca.mfaas.utils package
