+++ common-service-core/src/main/java/com/ca/mfaas/config/ApiInfo.java
@@ -9,7 +9,7 @@
  */
 package com.ca.mfaas.config;
 
-import com.ca.mfaas.eurekaservice.client.util.UrlUtils;
+import com.ca.mfaas.utils.UrlUtils;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
