+++ common-service-core/src/test/java/com/ca/mfaas/eurekaservice/client/util/StringUtilsTest.java
@@ -9,6 +9,7 @@
  */
 package com.ca.mfaas.eurekaservice.client.util;
 
+import com.ca.mfaas.utils.StringUtils;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
